refactor: move admin/core under admin/backend/core
Groups the editor's server-side workspace logic with the rest of the
backend package instead of a sibling top-level dir.
