+++ admin/backend/core/editor/search.py
